+++ docs/source/history.rst
@@ -18,6 +18,8 @@ Features
 
 - Convert to use Sphinx domains. Add ``spelling:word-list``
   directive. Have ``spelling`` directive report that it is deprecated.
+- Add ``spelling:word`` role for marking inline text as spelled
+  correctly.
 
 7.5.1
 =====